Multiplication result converted to larger type.

+++ src/generator_packed_gemm_bc_rm_aarch64.c
@@ -135,10 +135,10 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64( libxsmm_generated_code*       
     /* reset B and C pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    l_gp_reg_mapping.gp_reg_b, l_gp_reg_mapping.gp_reg_help_1, l_gp_reg_mapping.gp_reg_b,
-                                                   i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_in );
+                                                   (unsigned long long)i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_in );
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    l_gp_reg_mapping.gp_reg_c, l_gp_reg_mapping.gp_reg_help_2, l_gp_reg_mapping.gp_reg_c,
-                                                   i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_out );
+                                                   (unsigned long long)i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_out );
   } else if ( (l_n1_range > 0) && (l_n2_range == 0) ) {
     /* reset n loop */
     /* we have only one range */
@@ -152,10 +152,10 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64( libxsmm_generated_code*       
     /* reset B and C pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    l_gp_reg_mapping.gp_reg_b, l_gp_reg_mapping.gp_reg_help_1, l_gp_reg_mapping.gp_reg_b,
-                                                   i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_in );
+                                                   (unsigned long long)i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_in );
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    l_gp_reg_mapping.gp_reg_c, l_gp_reg_mapping.gp_reg_help_2, l_gp_reg_mapping.gp_reg_c,
-                                                   i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_out );
+                                                   (unsigned long long)i_xgemm_desc->n * i_packed_width * l_micro_kernel_config.datatype_size_out );
   } else {
     LIBXSMM_HANDLE_ERROR( io_generated_code, LIBXSMM_ERR_N_BLOCK );
     return;
@@ -164,11 +164,11 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64( libxsmm_generated_code*       
   /* advance A pointer */
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                  l_gp_reg_mapping.gp_reg_a, l_gp_reg_mapping.gp_reg_help_0, l_gp_reg_mapping.gp_reg_a,
-                                                 l_micro_kernel_config.datatype_size_in*i_xgemm_desc->lda );
+                                                 (unsigned long long)l_micro_kernel_config.datatype_size_in*i_xgemm_desc->lda );
   /* advance C pointer */
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                  l_gp_reg_mapping.gp_reg_c, l_gp_reg_mapping.gp_reg_help_2, l_gp_reg_mapping.gp_reg_c,
-                                                 l_micro_kernel_config.datatype_size_out*i_packed_width*i_xgemm_desc->ldc );
+                                                 (unsigned long long)l_micro_kernel_config.datatype_size_out*i_packed_width*i_xgemm_desc->ldc );
 
   /* close m loop */
   libxsmm_generator_loop_footer_aarch64( io_generated_code, &l_loop_label_tracker, l_gp_reg_mapping.gp_reg_mloop, 1 );
@@ -243,12 +243,12 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop( libxsmm_generated_code* 
     /* advance B pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                    i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_b,
-                                                   l_simd_packed_width * i_micro_kernel_config->datatype_size_in );
+                                                   (unsigned long long)l_simd_packed_width * i_micro_kernel_config->datatype_size_in );
 
     /* advance C pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                    i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                   l_simd_packed_width * i_micro_kernel_config->datatype_size_out );
+                                                   (unsigned long long)l_simd_packed_width * i_micro_kernel_config->datatype_size_out );
 
     /* jump back to pack loop label */
     libxsmm_generator_loop_footer_aarch64( io_generated_code, io_loop_label_tracker, i_gp_reg_mapping->gp_reg_help_3, l_simd_packed_width );
@@ -256,12 +256,12 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop( libxsmm_generated_code* 
     /* reset B pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_b,
-                                                   l_simd_packed_iters * l_simd_packed_width * i_micro_kernel_config->datatype_size_in );
+                                                   (unsigned long long)l_simd_packed_iters * l_simd_packed_width * i_micro_kernel_config->datatype_size_in );
 
     /* reset C pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                   l_simd_packed_iters * l_simd_packed_width * i_micro_kernel_config->datatype_size_out );
+                                                   (unsigned long long)l_simd_packed_iters * l_simd_packed_width * i_micro_kernel_config->datatype_size_out );
 
   /* we need masking and have less than SIMD width */
   } else if ( l_simd_packed_width > i_packed_width  ) {
@@ -276,12 +276,12 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop( libxsmm_generated_code* 
     /* advance B pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                    i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_b,
-                                                   i_n_blocking * i_packed_width * i_micro_kernel_config->datatype_size_in );
+                                                   (unsigned long long)i_n_blocking * i_packed_width * i_micro_kernel_config->datatype_size_in );
 
     /* advance C pointer */
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_ADD,
                                                    i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                   i_n_blocking * i_packed_width * i_micro_kernel_config->datatype_size_out );
+                                                   (unsigned long long)i_n_blocking * i_packed_width * i_micro_kernel_config->datatype_size_out );
   }
 }
 
@@ -325,7 +325,7 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop_simd_packed_asimd( libxsm
   if (0 == (LIBXSMM_GEMM_FLAG_BETA_0 & i_xgemm_desc->flags)) {
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                   i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
+                                                   (unsigned long long)i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
   }
 
   /* k loop */
@@ -373,17 +373,17 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop_simd_packed_asimd( libxsm
   }
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                  i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                 i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
+                                                 (unsigned long long)i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
 
   /* reset A pointer */
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                  i_gp_reg_mapping->gp_reg_a, i_gp_reg_mapping->gp_reg_help_0, i_gp_reg_mapping->gp_reg_a,
-                                                 (unsigned long long)((unsigned long long)i_xgemm_desc->k*i_micro_kernel_config->datatype_size_in) );
+                                                 (unsigned long long)i_xgemm_desc->k*i_micro_kernel_config->datatype_size_in );
 
   /* reset B pointer */
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                  i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_b,
-                                                 (unsigned long long)((unsigned long long)i_xgemm_desc->k * i_xgemm_desc->ldb * i_packed_width * i_micro_kernel_config->datatype_size_in) );
+                                                 (unsigned long long)i_xgemm_desc->k * i_xgemm_desc->ldb * i_packed_width * i_micro_kernel_config->datatype_size_in );
 }
 
 LIBXSMM_API_INTERN
@@ -432,7 +432,7 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop_simd_packed_sve( libxsmm_
   if (0 == (LIBXSMM_GEMM_FLAG_BETA_0 & i_xgemm_desc->flags)) {
     libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                    i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                   i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
+                                                   (unsigned long long)i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
   }
 
   /* k loop */
@@ -500,7 +500,7 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop_simd_packed_sve( libxsmm_
   }
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
                                                  i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, i_gp_reg_mapping->gp_reg_c,
-                                                 i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
+                                                 (unsigned long long)i_n_blocking*i_packed_width*i_micro_kernel_config->datatype_size_out );
 
   /* reset A pointer */
   libxsmm_aarch64_instruction_alu_compute_imm64( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_META_SUB,
@@ -512,4 +512,3 @@ void libxsmm_generator_packed_gemm_bc_rm_aarch64_kloop_simd_packed_sve( libxsmm_
                                                  i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_b,
                                                  (unsigned long long)((unsigned long long)i_xgemm_desc->k * i_xgemm_desc->ldb * i_packed_width * i_micro_kernel_config->datatype_size_in) );
 }
-
