+++ src/generator_packed_spgemm_csc_bsparse_aarch64.c
@@ -306,8 +306,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_asimd( libxsmm_ge
           /* @TODO */
         } else {
           libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_2,
-                                                     ( (i_n_processed + l_n)*i_packed_width*i_micro_kernel_config->datatype_size_out ) +
-                                                     ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_out ) );
+                                                     ( ((unsigned long long)i_n_processed + l_n)*i_packed_width*i_micro_kernel_config->datatype_size_out ) +
+                                                     ( ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_out ) );
           libxsmm_aarch64_instruction_asimd_move( io_generated_code, LIBXSMM_AARCH64_INSTR_ASIMD_LDR_R,
                                                   i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, 0,
                                                   l_reg, LIBXSMM_AARCH64_ASIMD_WIDTH_Q );
@@ -350,8 +350,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_asimd( libxsmm_ge
           /* @TODO */
         } else {
           libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_0,
-                                                     (l_k*i_packed_width*i_micro_kernel_config->datatype_size_in) +
-                                                     ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_in ) );
+                                                     ((unsigned long long)l_k*i_packed_width*i_micro_kernel_config->datatype_size_in) +
+                                                     ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_in );
           libxsmm_aarch64_instruction_asimd_move( io_generated_code, LIBXSMM_AARCH64_INSTR_ASIMD_LDR_R,
                                                   i_gp_reg_mapping->gp_reg_a, i_gp_reg_mapping->gp_reg_help_0, 0,
                                                   l_max_reg_block, LIBXSMM_AARCH64_ASIMD_WIDTH_Q );
@@ -360,7 +360,7 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_asimd( libxsmm_ge
         for ( l_n = 0; l_n < l_n_blocking; l_n++ ) {
           if ( l_nnz_idx[l_n][0] != -1 ) {
             libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_1,
-                                                       l_nnz_idx[l_n][0] * i_micro_kernel_config->datatype_size_in );
+                                                       (unsigned long long)l_nnz_idx[l_n][0] * i_micro_kernel_config->datatype_size_in );
             libxsmm_aarch64_instruction_asimd_move( io_generated_code, LIBXSMM_AARCH64_INSTR_ASIMD_LDR_R,
                                                     i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, 0,
                                                     l_max_reg_block+1, (i_micro_kernel_config->datatype_size_in == 4) ?  LIBXSMM_AARCH64_ASIMD_WIDTH_S : LIBXSMM_AARCH64_ASIMD_WIDTH_D );
@@ -383,8 +383,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_asimd( libxsmm_ge
       } else {
         unsigned int l_reg = (l_n*i_packed_blocking) + l_p;
         libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_2,
-                                                   ( (i_n_processed + l_n)*i_packed_width*i_micro_kernel_config->datatype_size_out ) +
-                                                   ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_out ) );
+                                                   ( ((unsigned long long)i_n_processed + l_n)*i_packed_width*i_micro_kernel_config->datatype_size_out ) +
+                                                   ( ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length*i_micro_kernel_config->datatype_size_out ) );
         libxsmm_aarch64_instruction_asimd_move( io_generated_code, LIBXSMM_AARCH64_INSTR_ASIMD_STR_R,
                                                 i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, 0,
                                                 l_reg, LIBXSMM_AARCH64_ASIMD_WIDTH_Q );
@@ -467,8 +467,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_sve( libxsmm_gene
           /* @TODO */
         } else {
           libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_2,
-                                                     ( (i_n_processed + l_n)*i_packed_width ) +
-                                                     ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length ) );
+                                                     ( ((unsigned long long)i_n_processed + l_n)*i_packed_width ) +
+                                                     ( ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length ) );
           libxsmm_aarch64_instruction_sve_move( io_generated_code,
                                                 (i_micro_kernel_config->datatype_size_out == 4) ? LIBXSMM_AARCH64_INSTR_SVE_LD1W_SR : LIBXSMM_AARCH64_INSTR_SVE_LD1D_SR,
                                                 i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, 0,
@@ -512,8 +512,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_sve( libxsmm_gene
           /* @TODO */
         } else {
           libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_0,
-                                                     (l_k*i_packed_width) +
-                                                     ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length ) );
+                                                     ((unsigned long long)l_k*i_packed_width) +
+                                                     ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length );
           libxsmm_aarch64_instruction_sve_move( io_generated_code,
                                                 (i_micro_kernel_config->datatype_size_in == 4) ? LIBXSMM_AARCH64_INSTR_SVE_LD1W_SR : LIBXSMM_AARCH64_INSTR_SVE_LD1D_SR,
                                                 i_gp_reg_mapping->gp_reg_a, i_gp_reg_mapping->gp_reg_help_0, 0,
@@ -523,7 +523,7 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_sve( libxsmm_gene
         for ( l_n = 0; l_n < l_n_blocking; l_n++ ) {
           if ( l_nnz_idx[l_n][0] != -1 ) {
             libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_1,
-                                                       l_nnz_idx[l_n][0] * i_micro_kernel_config->datatype_size_in );
+                                                       (unsigned long long)l_nnz_idx[l_n][0] * i_micro_kernel_config->datatype_size_in );
             libxsmm_aarch64_instruction_alu_compute_shifted_reg( io_generated_code, LIBXSMM_AARCH64_INSTR_GP_ADD_SR,
                                                                  i_gp_reg_mapping->gp_reg_b, i_gp_reg_mapping->gp_reg_help_1, i_gp_reg_mapping->gp_reg_help_1, 0, LIBXSMM_AARCH64_SHIFTMODE_LSL );
             libxsmm_aarch64_instruction_sve_move( io_generated_code,
@@ -552,8 +552,8 @@ void libxsmm_generator_packed_spgemm_csc_bsparse_aarch64_kloop_sve( libxsmm_gene
       } else {
         unsigned int l_reg = (l_n*i_packed_blocking) + l_p;
         libxsmm_aarch64_instruction_alu_set_imm64( io_generated_code, i_gp_reg_mapping->gp_reg_help_2,
-                                                   ( (i_n_processed + l_n)*i_packed_width ) +
-                                                   ( (i_packed_processed + l_p)*i_micro_kernel_config->vector_length ) );
+                                                   ( ((unsigned long long)i_n_processed + l_n)*i_packed_width ) +
+                                                   ( ((unsigned long long)i_packed_processed + l_p)*i_micro_kernel_config->vector_length ) );
         libxsmm_aarch64_instruction_sve_move( io_generated_code,
                                               (i_micro_kernel_config->datatype_size_out == 4) ? LIBXSMM_AARCH64_INSTR_SVE_ST1W_SR : LIBXSMM_AARCH64_INSTR_SVE_ST1D_SR,
                                               i_gp_reg_mapping->gp_reg_c, i_gp_reg_mapping->gp_reg_help_2, 0,
